+++ src/app/backend/auth/lib.ts
@@ -4,6 +4,9 @@ import { JWT } from "@auth/core/jwt";
 import { decodeJwt } from "jose";
 import { UserRole } from "@/app/models/entities/users";
 import { isAuthDisabledServerCheck } from "@/app/utils/auth";
+import KeycloakProvider from "next-auth/providers/keycloak";
+import MicrosoftEntraID from "next-auth/providers/microsoft-entra-id";
+import { Provider } from "@auth/core/providers";
 
 export interface AuthStrategy {
   parseIdpResponseForUserToken(
@@ -11,6 +14,7 @@ export interface AuthStrategy {
     account: Account,
     profile: Profile,
   ): UserToken;
+  setUpNextAuthProvider(): Provider;
 }
 
 const ROLE_TO_ENUM_MAP: Record<string, UserRole> = {
@@ -110,6 +114,39 @@ export class KeycloakAuthStrategy implements AuthStrategy {
     };
     return userToken;
   }
+  public setUpNextAuthProvider() {
+    function addRealm(url: string) {
+      return url.endsWith("/realms/master") ? url : `${url}/realms/master`;
+    }
+
+    let { NAMED_KEYCLOAK, LOCAL_KEYCLOAK } = process.env;
+    if (!NAMED_KEYCLOAK || !LOCAL_KEYCLOAK) {
+      const KEYCLOAK_URL =
+        process.env.AUTH_KEYCLOAK_ISSUER || "http://localhost:8080";
+      NAMED_KEYCLOAK = KEYCLOAK_URL;
+      LOCAL_KEYCLOAK = KEYCLOAK_URL;
+    }
+
+    // Add /realms/master to the end of the URL if it's missing.
+    NAMED_KEYCLOAK = addRealm(NAMED_KEYCLOAK);
+    LOCAL_KEYCLOAK = addRealm(LOCAL_KEYCLOAK);
+
+    return KeycloakProvider({
+      jwks_endpoint: `${NAMED_KEYCLOAK}/protocol/openid-connect/certs`,
+      wellKnown: undefined,
+      clientId: process.env.AUTH_CLIENT_ID,
+      clientSecret: process.env.AUTH_CLIENT_SECRET,
+      issuer: `${LOCAL_KEYCLOAK}`,
+      authorization: {
+        params: {
+          scope: "openid email profile",
+        },
+        url: `${LOCAL_KEYCLOAK}/protocol/openid-connect/auth`,
+      },
+      token: `${NAMED_KEYCLOAK}/protocol/openid-connect/token`,
+      userinfo: `${NAMED_KEYCLOAK}/protocol/openid-connect/userinfo`,
+    });
+  }
 }
 
 export class MicrosoftEntraAuthStrategy implements AuthStrategy {
@@ -135,5 +172,23 @@ export class MicrosoftEntraAuthStrategy implements AuthStrategy {
     };
 
     return userToken;
+  }
+
+  public setUpNextAuthProvider() {
+    return MicrosoftEntraID({
+      clientId: process.env.AUTH_CLIENT_ID,
+      clientSecret: process.env.AUTH_CLIENT_SECRET,
+      issuer: process.env.AUTH_ISSUER,
+      authorization: {
+        params: {
+          scope: "openid email profile",
+          claims: {
+            id_token: {
+              roles: { essential: true },
+            },
+          },
+        },
+      },
+    });
   }
 }
